+++ app/client/scripts/libs/lazyload.ts
@@ -1,0 +1,8 @@
+/**
+ * @file Lazy load images
+ */
+const LazyLoad = require('vanilla-lazyload'); // tslint:disable-line
+
+new LazyLoad({
+  elements_selector: '.js-lazyload',
+});