+++ app/client/scripts/libs/webfont.ts
@@ -1,0 +1,10 @@
+/**
+ * @file Load web fonts
+ */
+import WebFont from 'webfontloader';
+
+WebFont.load({
+  google: {
+    families: ['Droid Sans'], // TODO : Specify load fonts
+  },
+});