+++ app/client/scripts/libs/window.ts
@@ -1,0 +1,18 @@
+/**
+ * @file Window definitions
+ */
+
+/**
+ * Window definition of variables and functions.
+ */
+// tslint:disable-next-line
+interface Window {
+
+  /**
+   * GTM data layer
+   */
+  dataLayer: object[];
+
+}
+
+declare var window: Window;